requirements.txt
pygame-ce>=2.5.4
pillow>=12.0.0
pygbag>=0.9.2
screeninfo>=0.8.1
pyinstaller>=6.16.0
platformdirs>=4.5.0